Farms api almost done

+++ src/main/java/io/farmx/repository/RoleRepository.java
@@ -6,6 +6,7 @@ import io.farmx.model.Role;
 
 import java.util.Optional;
 
-public interface RoleRepository extends CrudRepository<Role, Integer> {
+public interface RoleRepository extends CrudRepository<Role, Long> {
+
     Optional<Role> findByName(String name);
 }
